Database Changes Added

+++ DatabaseAccess/tblSupplierInvoice.cs
@@ -25,7 +25,7 @@ namespace DatabaseAccess
         public int CompanyID { get; set; }
         public int BranchID { get; set; }
         public int InvoiceNo { get; set; }
-        public int TotalAmount { get; set; }
+        public double TotalAmount { get; set; }
         public System.DateTime InvoiceDate { get; set; }
         public string Description { get; set; }
         public int UserID { get; set; }
